+++ orc8r/cloud/go/services/directoryd/servicers/internal/servicer.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package internal_servicers
+package internal
 
 import (
 	"context"
