+++ orc8r/cloud/go/services/directoryd/servicers/internal/servicer_test.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package internal_servicers_test
+package internal_test
 
 import (
 	"context"
@@ -21,7 +21,7 @@ import (
 
 	"magma/orc8r/cloud/go/blobstore"
 	"magma/orc8r/cloud/go/services/directoryd/protos"
-	"magma/orc8r/cloud/go/services/directoryd/servicers/internal_servicers"
+	internal_servicers "magma/orc8r/cloud/go/services/directoryd/servicers/internal"
 	"magma/orc8r/cloud/go/services/directoryd/storage"
 	stateTestInit "magma/orc8r/cloud/go/services/state/test_init"
 	"magma/orc8r/cloud/go/sqorc"
